pyproject.toml
[project]
name = "playmcp-viewer"
version = "0.1.0"
description = "PlayMCP Viewer."
readme = "README.md"
authors = [
    { name = "yeongro", email = "zeroro.yun@gmail.com" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastmcp>=2.14.1,<3",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-json-logger>=4.0.0",
    "pyyaml>=6.0.3",
]

[project.scripts]

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"
